requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
websockets==12.0
opencv-python==4.9.0.80
numpy==1.26.4
openai==1.63.2
python-jose[cryptography]==3.3.0
elevenlabs==1.51.0
requests==2.31.0
httpx>=0.24.1 